requirements.txt
# Core PDF Processing (Lightweight)
pypdf>=3.17.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pillow>=10.0.0

# NLP (Efficient)
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0

# Databases
neo4j>=5.14.0
qdrant-client>=1.7.0

# Multi-Vector Embeddings
voyageai>=0.2.0
openai>=1.0.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2

# LLM
anthropic>=0.18.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
aiohttp>=3.9.0
pyyaml>=6.0.1
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
